[lang] move v2 module back to origin
